+++ src/main/java/com/oss/beellage/calendar/collection/CalendarCollection.java
@@ -1,7 +1,9 @@
 package com.oss.beellage.calendar.collection;
 
-import com.oss.beellage.calendar.domain.Calendar;
+import com.oss.beellage.calendar.Calendar;
+
 import java.util.List;
+
 import lombok.Data;
 
 @Data
