+++ src/main/java/com/oss/beellage/calendar/repository/CalendarRepository.java
@@ -1,6 +1,6 @@
 package com.oss.beellage.calendar.repository;
 
-import com.oss.beellage.calendar.domain.Calendar;
+import com.oss.beellage.calendar.Calendar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CalendarRepository extends JpaRepository<Calendar, Long> {
