+++ src/main/java/com/oss/beellage/calendar/service/CalendarService.java
@@ -1,9 +1,11 @@
 package com.oss.beellage.calendar.service;
 
-import com.oss.beellage.calendar.domain.Calendar;
+import com.oss.beellage.calendar.Calendar;
 import com.oss.beellage.calendar.dto.CalendarRequest;
 import com.oss.beellage.calendar.repository.CalendarRepository;
+
 import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
