+++ src/main/java/com/oss/beellage/chat/Chat.java
@@ -1,6 +1,6 @@
 package com.oss.beellage.chat;
 
-import com.oss.beellage.team.domain.Team;
+import com.oss.beellage.team.Team;
 import com.oss.beellage.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
