+++ src/main/java/com/oss/beellage/issue/collection/IssueCollection.java
@@ -1,7 +1,9 @@
 package com.oss.beellage.issue.collection;
 
-import com.oss.beellage.issue.domain.Issue;
+import com.oss.beellage.issue.Issue;
+
 import java.util.List;
+
 import lombok.Data;
 
 @Data
