+++ src/main/java/com/oss/beellage/issue/repository/IssueRepository.java
@@ -1,6 +1,6 @@
 package com.oss.beellage.issue.repository;
 
-import com.oss.beellage.issue.domain.Issue;
+import com.oss.beellage.issue.Issue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IssueRepository extends JpaRepository<Issue, Long> {
