+++ src/main/java/com/oss/beellage/project/Project.java
@@ -1,32 +1,44 @@
-package com.oss.beellage.team.domain;
+package com.oss.beellage.project;
 
-import com.oss.beellage.calendar.domain.Calendar;
-import com.oss.beellage.project.domain.Project;
+import com.oss.beellage.issue.Issue;
+import com.oss.beellage.schedule.Schedule;
+import com.oss.beellage.team.Team;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+
 import java.sql.Timestamp;
 import java.util.List;
+
 import lombok.Data;
 
 @Entity
-@Table(name = "teams")
+@Table(name = "projects")
 @Data
-public class Team {
+public class Project {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @ManyToOne
+    @JoinColumn(name = "team_id", nullable = false)
+    private Team team;
+
+    @OneToMany(mappedBy = "project", cascade = CascadeType.ALL)
+    private List<Issue> issues;
+
+    @OneToMany(mappedBy = "project", cascade = CascadeType.ALL)
+    private List<Schedule> schedules;
+
     @Column(name = "name", nullable = false)
     private String name;
-
-    @Column(name = "host_id", nullable = false)
-    private Long hostId;
 
     @Column(name = "created_at", nullable = false)
     private Timestamp createdAt;
@@ -39,10 +51,4 @@ public class Team {
 
     @Column(name = "description")
     private String description;
-
-    @OneToMany(mappedBy = "team", cascade = CascadeType.ALL)
-    private List<Project> projects;
-
-    @OneToMany(mappedBy = "team", cascade = CascadeType.ALL)
-    private List<Calendar> calendars;
 }
