+++ src/main/java/com/oss/beellage/project/repository/ProjectRepository.java
@@ -1,6 +1,6 @@
 package com.oss.beellage.project.repository;
 
-import com.oss.beellage.project.domain.Project;
+import com.oss.beellage.project.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjectRepository extends JpaRepository<Project, Long> {
