+++ src/main/java/com/oss/beellage/schedule/Schedule.java
@@ -1,9 +1,9 @@
-package com.oss.beellage.schedule.domain;
+package com.oss.beellage.schedule;
 
 
-import com.oss.beellage.calendar.domain.Calendar;
-import com.oss.beellage.issue.domain.Issue;
-import com.oss.beellage.project.domain.Project;
+import com.oss.beellage.calendar.Calendar;
+import com.oss.beellage.issue.Issue;
+import com.oss.beellage.project.Project;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -13,7 +13,9 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
+
 import java.sql.Timestamp;
+
 import lombok.Data;
 
 @Entity
