+++ src/main/java/com/oss/beellage/schedule/collection/ScheduleCollection.java
@@ -1,7 +1,9 @@
 package com.oss.beellage.schedule.collection;
 
-import com.oss.beellage.schedule.domain.Schedule;
+import com.oss.beellage.schedule.Schedule;
+
 import java.util.List;
+
 import lombok.Data;
 
 @Data
