+++ src/main/java/com/oss/beellage/schedule/controller/ScheduleController.java
@@ -1,10 +1,12 @@
 package com.oss.beellage.schedule.controller;
 
-import com.oss.beellage.schedule.domain.Schedule;
+import com.oss.beellage.schedule.Schedule;
 import com.oss.beellage.schedule.dto.ScheduleRequest;
 import com.oss.beellage.schedule.service.ScheduleService;
+
 import java.util.List;
 import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
