+++ src/main/java/com/oss/beellage/team/Team.java
@@ -1,42 +1,34 @@
-package com.oss.beellage.project.domain;
+package com.oss.beellage.team;
 
-import com.oss.beellage.issue.domain.Issue;
-import com.oss.beellage.schedule.domain.Schedule;
-import com.oss.beellage.team.domain.Team;
+import com.oss.beellage.calendar.Calendar;
+import com.oss.beellage.project.Project;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+
 import java.sql.Timestamp;
 import java.util.List;
+
 import lombok.Data;
 
 @Entity
-@Table(name = "projects")
+@Table(name = "teams")
 @Data
-public class Project {
+public class Team {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
-    @JoinColumn(name = "team_id", nullable = false)
-    private Team team;
-
-    @OneToMany(mappedBy = "project", cascade = CascadeType.ALL)
-    private List<Issue> issues;
-
-    @OneToMany(mappedBy = "project", cascade = CascadeType.ALL)
-    private List<Schedule> schedules;
-
     @Column(name = "name", nullable = false)
     private String name;
+
+    @Column(name = "host_id", nullable = false)
+    private Long hostId;
 
     @Column(name = "created_at", nullable = false)
     private Timestamp createdAt;
@@ -49,4 +41,10 @@ public class Project {
 
     @Column(name = "description")
     private String description;
+
+    @OneToMany(mappedBy = "team", cascade = CascadeType.ALL)
+    private List<Project> projects;
+
+    @OneToMany(mappedBy = "team", cascade = CascadeType.ALL)
+    private List<Calendar> calendars;
 }
