+++ src/main/java/com/oss/beellage/team/repository/TeamRepository.java
@@ -1,6 +1,6 @@
 package com.oss.beellage.team.repository;
 
-import com.oss.beellage.team.domain.Team;
+import com.oss.beellage.team.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamRepository extends JpaRepository<Team, Long> {
