+++ src/test/java/com/oss/beellage/team/service/TeamServiceTests.java
@@ -5,10 +5,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import com.oss.beellage.team.domain.Team;
+import com.oss.beellage.team.Team;
 import com.oss.beellage.team.dto.TeamRequest;
 import com.oss.beellage.team.repository.TeamRepository;
+
 import java.util.Optional;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
